Fix Gravship Landing on Anchored Map
Flying back to an anchored map is handled differently.

Fixes:

Map could not be found when opening a session due to the marker being initialized differently.

Marker could not be synced correctly because it was not spawned or did not have a parent map.

Deselects the gravship designator if another client selects a landing position.