Refactor ThingsById system to use non static data
The static data causes problems when two saves are played in one process.

In the first game, many maps are used, so some things have a high mapIndex.

In the next game, fewer maps are used and one map is abandoned.

When the static data is not cleared correctly, and when map gets abandoned,

thingsById.RemoveAll(kv => kv.Value.Map == map);

calls a map getter that uses Find.Maps and tries to access it with the thing mapIndex from the old save, which leads to an out-of-bounds error.