Fix Lights Not Turning On Final Version
Fix lights not turning on during and after gravship landing animation

Previously, lights would fail to turn on after landing until you either changed their color or rejoined the game.

In our multiplayer layer, some systems are updated in Tick that in singleplayer are normally updated in UpdateLoop, not via the tick system.

This commit makes sure those systems are updated once after gravship things are spawned on the map.