Remove anomaly check from colony subhuman check
In 1.5, this used to be a check for colony mutant, and included Anomaly check since base game did so as well.

In 1.6, the code we patched was replaced to not be a mutant check, but a subhuman check. This check doesn't include a check for Anomaly DLC, so I've removed it to match base game behavior.