Fix Gravship Launch Anchored Map Session Creation
Launching a gravship from an anchored map failed to create the session correctly when the world tile picker is selected.

This bug was introduced by my recent refactor of GravshipTravelSessionPatches.cs.

The fix updates how preconfirm dialog options are synced, removing the need to sync `StartChoosingDestination_NewTemp`, where the postfix and `!Multiplayer.ExecutingCmds` check caused the issue.

Additionally, removed some code that was needed because the method was synced.