Remove Rbp Probing (Fixes #612, #658)
RBD probing was originally to address 1.6 issues with crossplay between Windows and Linux, as well as when players were not using the same build of the mod. This issue has now been resolved by #623.

The probing caused errors when a constructor was invoked via reflection. I don’t have an explanation for why this is happening, but it’s not worth investigating at the moment, as we are reverting the code to the state that Zetrith implemented, along with the #623 fix.