Change how Odyssey's `ShouldSpawnMotesAt` fixes are handled (#665)
The issue with those methods calling `ShouldSpawnMotesAt` is that they use RNG, changing the state of RNG depending on player-specific info (current map).

Currently, the methods spawning flecks are handled in OdysseyMultifaction class. Those patches ensure that the flecks will always spawn, regardless of the map the player is on.

However, we generally handle desyncs due to fleck/mote spawning like this using a prefix/finalizer and pushin/popping the RNG state, so even if the methods themselves use RNG depending on the player-specific info they'll all be safe.

Handling it by using `Rand.PushState` and `Rand.PopState` is generally going to be better performance-wise, since we won't be spawning motes/flecks on maps the players isn't currently looking at. On top of that, the patches themselves are also going to be much simpler now.

Changes:
- Added 3 Odyssey fleck methods using `ShouldSpawnMotesAt` to list of methods that are patched with Push/Pop State
- Removed OdysseyMultifaction.cs, as it's no longer needed