Optimize harmony metadata generation by generating patch info only once
Internally Harmony.VersionInfo iterates over all patched methods and
gets generates PatchInfo for each of them. This is something we already
do ourselves for describing the patched methods list. This probably
woudn't be much of a issue, however PatchInfos are stored at rest only
as a serialiazed binary blob, which means they all are deserialized when
generating the report. The base game with this mod has over 1000
patches which causes the whole process to be time consuming.