Trigger desync debug action, faster desync report generation, and separated metadata in the report (#667)
* Minor changes around SyncCoordinator

* Add debug action: trigger desync

* Move harmony metadata into it's own file in the desync report

It can easily generate over 150kB of data. Moving it out to a separate
file makes sense content-wise and makes it easier to browse the log file

* Optimize harmony metadata generation by generating patch info only once

Internally Harmony.VersionInfo iterates over all patched methods and
gets generates PatchInfo for each of them. This is something we already
do ourselves for describing the patched methods list. This probably
woudn't be much of a issue, however PatchInfos are stored at rest only
as a serialiazed binary blob, which means they all are deserialized when
generating the report. The base game with this mod has over 1000
patches which causes the whole process to be time consuming.

* Describe patched methods a bit faster

* Start generating desync metadata in the background

* Copy the pdb files during build

* Simplify MultiplayerLoader

* Avoid game freeze during desync report saving