Improve handling of packets when loading (#668)
* Only send Cursor packets to ingame players

As long as a player isn't fully loaded it doesn't make sense to send those packets.

* Handle TimeControl and KeepAlive packets during loading as a client

The server sends them regardless if the client is already playing or
just loading. While the server could be made to only send those packets
after the client has finished loading, the packets are generally simple
enough to be easily handled during loading too.

* Only send client cursor to server after loading is finished

It doesn't make sense to send cursor position to server when the client
might not even see the map yet